// frontend/src/app/core/api/generated/model/stock-movement-page-response.ts
/**
 * Inventory API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { StockMovementResponse } from './stock-movement-response';


export interface StockMovementPageResponse { 
    content?: Array<StockMovementResponse>;
    page?: number;
    size?: number;
    totalElements?: number;
    totalPages?: number;
    first?: boolean;
    last?: boolean;
}