// frontend/src/app/core/api/generated/model/stock-movement-response.ts
/**
 * Inventory API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface StockMovementResponse { 
    id?: string;
    warehouseId?: string;
    productId?: string;
    movementType?: StockMovementResponse.MovementTypeEnum;
    quantityDelta?: number;
    balanceBefore?: number;
    balanceAfter?: number;
    reservationDelta?: number;
    reservedBefore?: number;
    reservedAfter?: number;
    businessReference?: string;
    occurredAt?: string;
    responsibleUser?: string;
}
export namespace StockMovementResponse {
    export const MovementTypeEnum = {
        InitialStock: 'INITIAL_STOCK',
        ManualIn: 'MANUAL_IN',
        ManualOut: 'MANUAL_OUT',
        OrderReserved: 'ORDER_RESERVED',
        OrderReservationReleased: 'ORDER_RESERVATION_RELEASED',
        OrderConfirmed: 'ORDER_CONFIRMED',
        OrderCancelled: 'ORDER_CANCELLED',
        PurchaseReceived: 'PURCHASE_RECEIVED',
        TransferOut: 'TRANSFER_OUT',
        TransferIn: 'TRANSFER_IN',
        PhysicalCountAdjustment: 'PHYSICAL_COUNT_ADJUSTMENT'
    } as const;
    export type MovementTypeEnum = typeof MovementTypeEnum[keyof typeof MovementTypeEnum];
}