// frontend/src/app/core/api/generated/model/warehouse-response.ts
/**
 * Inventory API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface WarehouseResponse { 
    id?: string;
    code?: string;
    name?: string;
    description?: string;
    active?: boolean;
    createdAt?: string;
    updatedAt?: string;
}